unit test
wrap repetitive codes
remove test/unit folder
setup jest.config.js
