+++ src/components/Button/Button.spec.js
@@ -3,7 +3,7 @@ import Button from './Button.vue';
 
 describe('Button.vue', () => {
   it('renders props.label when passed', () => {
-    const msg = 'button test';
+    const msg = 'Button';
     const wrapper = shallowMount(Button, {
       props: { msg },
     });
